plug-ins: Don't use compose-type on Recompose
Resolves #12623
Compose and Recompose are in the same plug-in.
When we converted compose.c to use GimpChoice in 7b29ac24,
we did so without checking if it was being called by
Compose or Recompose. As a result, we always overwrote
compose_type with 0 (RGB mode) on Recompose since it
did not have a "compose-type" GimpChoice property.

This patch adds a check to make sure we're not in a Recompose
call before trying to set compose_type from the GimpChoice
parameter.