build/windows: Do not install 'lensfun' for obscure/broken gegl:lens-correct
It's very unclear who and when added that package to the deps list and if
it is needed to we bundle something more than the library (share/lensfun?).

Indeed, that GEGL op is reported to not work properly as a tool (2ba36733) and
have severe UX problems (#4695), being impossible to use even as experimental.
So, let's not install it to not bloat GIMP installation with not used code.