Revert "Issue #7539: Opening some images lock GIMP."
This reverts commit 9b94e347bcb39c1f5bed2f0df16cbd7deefa6ff4.

I leave a comment though, because we are not 100% sure that the issue is
gone. In fact, the upstream report is still opened.

The good point is that on stable code, we only debug crashes by default,
while the specific hang we had was happening with debugging metadata
warnings. So hopefully we won't have random people reporting GIMP
hanging (only people explicitly trying to help debugging GIMP).