libgimpcolor: GIMP built-in sRGB profiles now generated by babl.
Though it also seems like a next step towards babl-managed color
management, the first reason is that LittleCMS adds a creation
timestamp.
Looking up upstream code, it looks like there is no way to override this
value (Icc->Created is only set at creation and none of the time tags
seem to touch this).

Unfortunately it makes our built-in sRGB profiles non-reproducible
binaries (which for instance is a problem in gimp-date#7).