gui, themes: Recognize macOS 'Differentiate without Color' setting
macOS has a 'Differentiate without Color'
setting, which changes certain options to
be more distinct beyond color changes.
In GIMP, the main area seems to be the
GtkSwitches, which rely on color to show
if they're on or off. We have a CSS style
that hides I/O images to better match
2.10 style.

This patch moves that style from common.css
to /app/gu/themes, and conditions it on
the macOS setting. For other OS, the
style is added by default.