app: allow `gimp` command-line to run in console mode
When non-interactive mode is requested (the -i command-line flag),
_and_ we've compiled the app with UI support enabled, allow `gimp`
to branch into the console-based version of the app -- equivalent to
running it using `gimp-console`.

This avoids GTK-related failures open a display when one is not
not available (for example, when running `gimp` within a container
or a VM without a graphical display).

(cherry picked from commit ea0574b86c312f1d59050dbcdde36c2eb7031471)